Migrated the tests to TypeScript.
