+++ packages/lit-virtualizer/tsconfig.json
@@ -11,7 +11,7 @@
     "inlineSources": true,
     // "outDir": "./development/",
     "outDir": "./",
-    "rootDir": "./src",
+    // "rootDir": "./src",
     "strict": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
@@ -24,8 +24,9 @@
     "experimentalDecorators": true,
     "importHelpers": true
   },
-  "include": ["src/**/*.ts"],
-  "exclude": [],
+  "include": [
+    "src/**/*.ts"
+  ],
   "references": []
 }
 // {
